QWIZERO_WEB/requirements.txt
flask
gunicorn
Pillow